Moved Canvas rendering to a single thread.
