+++ components/canvas/lib.rs
@@ -20,6 +20,7 @@ extern crate servo_config;
 extern crate webrender;
 extern crate webrender_api;
 
+pub mod canvas_data;
 pub mod canvas_paint_thread;
 pub mod gl_context;
 mod webgl_mode;
